update token fetching control flow to avoid stale build state
