+++ server/internal/api/infrastructure/registry/client.go
@@ -20,21 +20,6 @@ type Client struct {
 	httpClient *http.Client
 }
 
-type staticCredentialStore struct {
-	username string
-	password string
-}
-
-func (s *staticCredentialStore) Basic(*url.URL) (string, string) {
-	return s.username, s.password
-}
-
-func (s *staticCredentialStore) RefreshToken(*url.URL, string) string {
-	return ""
-}
-
-func (s *staticCredentialStore) SetRefreshToken(*url.URL, string, string) {}
-
 func NewClient(cfg *conf.Config) port.Registry {
 	transport := http.DefaultTransport.(*http.Transport).Clone()
 	transport.ResponseHeaderTimeout = 10 * time.Second
